Renamed createUserFolder.ts
