+++ server/src/routes/auth-routes.ts
@@ -2,7 +2,7 @@ import { Router, Request, Response } from "express";
 import { User } from "../models/user";
 import jwt from "jsonwebtoken";
 import bcrypt from "bcrypt";
-import { createUserFolder } from "../utils/createUserFolder";
+import { createUserFolder } from "../utils/userFolder";
 
 export const login = async (req: Request, res: any) => {
   const { email, password } = req.body;
